SP-3: Implement CommentController and its supporting methods in CommentServiceImpl and CommentRepositoryImpl

+++ src/main/java/org/practicum/yandex/controller/PostController.java
@@ -119,7 +119,7 @@ public class PostController {
         } catch (DataNotFoundException e) {
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND,
-                    String.format("Post with id %d not found", postId.longValue())
+                    String.format("Post with id %d not found", postId)
             );
         }
     }
